+++ libs/backpack-adoption-action/src/index.ts
@@ -1,4 +1,4 @@
-import { run } from "./action";
+import { run } from "./action/run";
 
 run().catch((error: unknown) => {
   const message = error instanceof Error ? error.message : String(error);
